moved GetAttorneyStatus
