+++ service-api/app/src/App/src/Service/Lpa/SiriusLpaManager.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Service\Lpa;
 
-use App\Service\Features\FeatureEnabled;
 use App\DataAccess\Repository\{InstructionsAndPreferencesImagesInterface,
     LpasInterface,
     Response\Lpa,
@@ -13,6 +12,8 @@ use App\DataAccess\Repository\{InstructionsAndPreferencesImagesInterface,
     ViewerCodeActivityInterface,
     ViewerCodesInterface};
 use App\Exception\GoneException;
+use App\Service\Features\FeatureEnabled;
+use App\Service\Lpa\GetAttorneyStatus\GetAttorneyStatus;
 use DateTime;
 use Psr\Log\LoggerInterface;
 use RuntimeException;
